update from origin

+++ src/vstart.sh
@@ -411,6 +411,10 @@ prepare_conf() {
         plugin dir = $CEPH_LIB
         osd pool default erasure code profile = plugin=jerasure technique=reed_sol_van k=2 m=1 ruleset-failure-domain=osd
         rgw frontends = $rgw_frontend port=$CEPH_RGW_PORT
+        ; needed for s3tests
+        rgw crypt s3 kms encryption keys = testkey-1=YmluCmJvb3N0CmJvb3N0LWJ1aWxkCmNlcGguY29uZgo= testkey-2=aWIKTWFrZWZpbGUKbWFuCm91dApzcmMKVGVzdGluZwo=
+        rgw crypt require ssl = false
+        rgw lc debug interval = 10
         filestore fd cache size = 32
         run dir = $CEPH_OUT_DIR
         enable experimental unrecoverable data corrupting features = *
@@ -495,7 +499,6 @@ $extra_conf
 [mon]
         mon pg warn min per osd = 3
         mon osd allow primary affinity = true
-        mon osd allow pg upmap = true
         mon reweight min pgs per osd = 4
         mon osd prime pg temp = true
         crushtool = $CEPH_BIN/crushtool
@@ -634,16 +637,8 @@ EOF
 	DASH_URLS+="http://$IP:$MGR_PORT/"
 	MGR_PORT=$(($MGR_PORT + 1000))
 
-	CERT=`mktemp`
-	PKEY=`mktemp`
-	openssl req -new -nodes -x509 \
-		-subj "/O=IT/CN=ceph-mgr-restful" \
-		-days 3650 -keyout "$PKEY" -out "$CERT" -extensions v3_ca
 	ceph_adm config-key put mgr/restful/$name/server_addr $IP
 	ceph_adm config-key put mgr/restful/$name/server_port $MGR_PORT
-	ceph_adm config-key put mgr/restful/$name/crt -i $CERT
-	ceph_adm config-key put mgr/restful/$name/key -i $PKEY
-	rm $CERT $PKEY
 
 	RESTFUL_URLS+="https://$IP:$MGR_PORT"
 	MGR_PORT=$(($MGR_PORT + 1000))
@@ -652,10 +647,14 @@ EOF
         run 'mgr' $CEPH_BIN/ceph-mgr -i $name $ARGS
     done
 
-    SF=`mktemp`
-    ceph_adm tell mgr.x restful create-key admin -o $SF
-    RESTFUL_SECRET=`cat $SF`
-    rm $SF
+    if ceph_adm tell mgr restful create-self-signed-cert; then
+        SF=`mktemp`
+        ceph_adm tell mgr restful create-key admin -o $SF
+        RESTFUL_SECRET=`cat $SF`
+        rm $SF
+    else 
+        echo MGR Restful is not working, perhaps the package is not installed?
+    fi
 }
 
 start_mds() {
